api/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
supabase>=2.28.0
pyjwt>=2.9.0
python-dotenv>=1.0.0
httpx>=0.28.0
anthropic>=0.52.0
